Clarify re-key readme and split EET helper into two suites
